+++ modules/openapi-generator/src/test/resources/3_0/python-experimental/petstore-with-fake-endpoints-models-for-testing-with-http-signature.yaml
@@ -1407,6 +1407,7 @@ components:
         name: Tag
     Pet:
       type: object
+      description: Pet object that needs to be added to the store
       required:
         - name
         - photoUrls
