+++ samples/openapi3/client/petstore/python-experimental/.openapi-generator/FILES
@@ -209,4 +209,99 @@ setup.cfg
 setup.py
 test-requirements.txt
 test/__init__.py
+test/test_additional_properties_class.py
+test/test_additional_properties_with_array_of_enums.py
+test/test_address.py
+test/test_animal.py
+test/test_animal_farm.py
+test/test_another_fake_api.py
+test/test_api_response.py
+test/test_apple.py
+test/test_apple_req.py
+test/test_array_of_array_of_number_only.py
+test/test_array_of_enums.py
+test/test_array_of_number_only.py
+test/test_array_test.py
+test/test_banana.py
+test/test_banana_req.py
+test/test_basque_pig.py
+test/test_capitalization.py
+test/test_cat.py
+test/test_cat_all_of.py
+test/test_category.py
+test/test_child_cat.py
+test/test_child_cat_all_of.py
+test/test_class_model.py
+test/test_client.py
+test/test_complex_quadrilateral.py
+test/test_composed_one_of_number_with_validations.py
+test/test_danish_pig.py
+test/test_default_api.py
+test/test_dog.py
+test/test_dog_all_of.py
+test/test_drawing.py
+test/test_enum_arrays.py
+test/test_enum_class.py
+test/test_enum_test.py
+test/test_equilateral_triangle.py
+test/test_fake_api.py
+test/test_fake_classname_tags_123_api.py
+test/test_file.py
+test/test_file_schema_test_class.py
+test/test_foo.py
+test/test_format_test.py
+test/test_fruit.py
+test/test_fruit_req.py
+test/test_gm_fruit.py
+test/test_grandparent_animal.py
+test/test_has_only_read_only.py
+test/test_health_check_result.py
+test/test_inline_object.py
+test/test_inline_object1.py
+test/test_inline_object2.py
+test/test_inline_object3.py
+test/test_inline_object4.py
+test/test_inline_object5.py
+test/test_inline_response_default.py
+test/test_integer_enum.py
+test/test_integer_enum_one_value.py
+test/test_integer_enum_with_default_value.py
+test/test_isosceles_triangle.py
+test/test_list.py
+test/test_mammal.py
+test/test_map_test.py
+test/test_mixed_properties_and_additional_properties_class.py
+test/test_model200_response.py
+test/test_model_return.py
+test/test_name.py
+test/test_nullable_class.py
+test/test_nullable_shape.py
+test/test_number_only.py
+test/test_number_with_validations.py
+test/test_object_model_with_ref_props.py
+test/test_order.py
+test/test_parent_pet.py
+test/test_pet.py
+test/test_pet_api.py
+test/test_pig.py
+test/test_quadrilateral.py
+test/test_quadrilateral_interface.py
+test/test_read_only_first.py
+test/test_scalene_triangle.py
+test/test_shape.py
+test/test_shape_interface.py
+test/test_shape_or_null.py
+test/test_simple_quadrilateral.py
+test/test_special_model_name.py
+test/test_store_api.py
+test/test_string_boolean_map.py
+test/test_string_enum.py
+test/test_string_enum_with_default_value.py
+test/test_tag.py
+test/test_triangle.py
+test/test_triangle_interface.py
+test/test_user.py
+test/test_user_api.py
+test/test_whale.py
+test/test_zebra.py
 tox.ini
