+++ samples/openapi3/client/petstore/python-experimental/docs/Pet.md
@@ -1,5 +1,6 @@
 # Pet
 
+Pet object that needs to be added to the store
 ## Properties
 Name | Type | Description | Notes
 ------------ | ------------- | ------------- | -------------
