+++ samples/openapi3/client/petstore/python-experimental/test/test_additional_properties_with_array_of_enums.py
@@ -14,11 +14,8 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import enum_class
-except ImportError:
-    enum_class = sys.modules[
-        'petstore_api.model.enum_class']
+from petstore_api.model.enum_class import EnumClass
+globals()['EnumClass'] = EnumClass
 from petstore_api.model.additional_properties_with_array_of_enums import AdditionalPropertiesWithArrayOfEnums
