+++ samples/openapi3/client/petstore/python-experimental/test/test_animal_farm.py
@@ -14,11 +14,8 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import animal
-except ImportError:
-    animal = sys.modules[
-        'petstore_api.model.animal']
+from petstore_api.model.animal import Animal
+globals()['Animal'] = Animal
 from petstore_api.model.animal_farm import AnimalFarm
