+++ samples/openapi3/client/petstore/python-experimental/test/test_cat.py
@@ -14,21 +14,12 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import address
-except ImportError:
-    address = sys.modules[
-        'petstore_api.model.address']
-try:
-    from petstore_api.model import animal
-except ImportError:
-    animal = sys.modules[
-        'petstore_api.model.animal']
-try:
-    from petstore_api.model import cat_all_of
-except ImportError:
-    cat_all_of = sys.modules[
-        'petstore_api.model.cat_all_of']
+from petstore_api.model.address import Address
+from petstore_api.model.animal import Animal
+from petstore_api.model.cat_all_of import CatAllOf
+globals()['Address'] = Address
+globals()['Animal'] = Animal
+globals()['CatAllOf'] = CatAllOf
 from petstore_api.model.cat import Cat
