+++ samples/openapi3/client/petstore/python-experimental/test/test_child_cat.py
@@ -14,16 +14,10 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import child_cat_all_of
-except ImportError:
-    child_cat_all_of = sys.modules[
-        'petstore_api.model.child_cat_all_of']
-try:
-    from petstore_api.model import parent_pet
-except ImportError:
-    parent_pet = sys.modules[
-        'petstore_api.model.parent_pet']
+from petstore_api.model.child_cat_all_of import ChildCatAllOf
+from petstore_api.model.parent_pet import ParentPet
+globals()['ChildCatAllOf'] = ChildCatAllOf
+globals()['ParentPet'] = ParentPet
 from petstore_api.model.child_cat import ChildCat
