+++ samples/openapi3/client/petstore/python-experimental/test/test_file_schema_test_class.py
@@ -14,11 +14,8 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import file
-except ImportError:
-    file = sys.modules[
-        'petstore_api.model.file']
+from petstore_api.model.file import File
+globals()['File'] = File
 from petstore_api.model.file_schema_test_class import FileSchemaTestClass
