+++ samples/openapi3/client/petstore/python-experimental/test/test_gm_fruit.py
@@ -14,16 +14,10 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import apple
-except ImportError:
-    apple = sys.modules[
-        'petstore_api.model.apple']
-try:
-    from petstore_api.model import banana
-except ImportError:
-    banana = sys.modules[
-        'petstore_api.model.banana']
+from petstore_api.model.apple import Apple
+from petstore_api.model.banana import Banana
+globals()['Apple'] = Apple
+globals()['Banana'] = Banana
 from petstore_api.model.gm_fruit import GmFruit
 
 
@@ -38,193 +32,10 @@ class TestGmFruit(unittest.TestCase):
 
     def testGmFruit(self):
         """Test GmFruit"""
+        # FIXME: construct object with mandatory attributes with example values
+        # model = GmFruit()  # noqa: E501
+        pass
 
-        # make an instance of GmFruit, a composed schema anyOf model
-        # banana test
-        length_cm = 20.3
-        color = 'yellow'
-        fruit = GmFruit(length_cm=length_cm, color=color)
-        # check its properties
-        self.assertEqual(fruit.length_cm, length_cm)
-        self.assertEqual(fruit['length_cm'], length_cm)
-        self.assertEqual(getattr(fruit, 'length_cm'), length_cm)
-        self.assertEqual(fruit.color, color)
-        self.assertEqual(fruit['color'], color)
-        self.assertEqual(getattr(fruit, 'color'), color)
-        # check the dict representation
-        self.assertEqual(
-            fruit.to_dict(),
-            {
-                'length_cm': length_cm,
-                'color': color
-            }
-        )
-        # setting a value that doesn't exist raises an exception
-        # with a key
-        with self.assertRaises(AttributeError):
-            fruit['invalid_variable'] = 'some value'
-        # with setattr
-        with self.assertRaises(AttributeError):
-            setattr(fruit, 'invalid_variable', 'some value')
-
-        # getting a value that doesn't exist raises an exception
-        # with a key
-        with self.assertRaises(AttributeError):
-            invalid_variable = fruit['cultivar']
-        # with getattr
-        self.assertTrue(getattr(fruit, 'cultivar', 'some value'), 'some value')
-
-        with self.assertRaises(AttributeError):
-            invalid_variable = getattr(fruit, 'cultivar')
-
-        # make sure that the ModelComposed class properties are correct
-        # model._composed_schemas stores the anyOf/allOf/oneOf info
-        self.assertEqual(
-            fruit._composed_schemas,
-            {
-                'anyOf': [
-                    apple.Apple,
-                    banana.Banana,
-                ],
-                'allOf': [],
-                'oneOf': [],
-            }
-        )
-        # model._composed_instances is a list of the instances that were
-        # made from the anyOf/allOf/OneOf classes in model._composed_schemas
-        for composed_instance in fruit._composed_instances:
-            if composed_instance.__class__ == banana.Banana:
-                banana_instance = composed_instance
-        self.assertEqual(
-            fruit._composed_instances,
-            [banana_instance]
-        )
-        # model._var_name_to_model_instances maps the variable name to the
-        # model instances which store that variable
-        self.assertEqual(
-            fruit._var_name_to_model_instances,
-            {
-                'color': [fruit],
-                'length_cm': [fruit, banana_instance],
-                'cultivar': [fruit],
-                'origin': [fruit],
-            }
-        )
-        # model._additional_properties_model_instances stores a list of
-        # models which have the property additional_properties_type != None
-        self.assertEqual(
-            fruit._additional_properties_model_instances, []
-        )
-
-        # if we modify one of the properties owned by multiple
-        # model_instances we get an exception when we try to access that
-        # property because the retrieved values are not all the same
-        banana_instance.length_cm = 4.56
-        with self.assertRaises(petstore_api.ApiValueError):
-            some_length_cm = fruit.length_cm
-
-        # including extra parameters raises an exception
-        with self.assertRaises(petstore_api.ApiValueError):
-            fruit = GmFruit(
-                color=color,
-                length_cm=length_cm,
-                unknown_property='some value'
-            )
-
-        # including input parameters for both anyOf instances works
-        cultivar = 'banaple'
-        color = 'orange'
-        fruit = GmFruit(
-            color=color,
-            cultivar=cultivar,
-            length_cm=length_cm
-        )
-        self.assertEqual(fruit.color, color)
-        self.assertEqual(fruit['color'], color)
-        self.assertEqual(getattr(fruit, 'color'), color)
-        self.assertEqual(fruit.cultivar, cultivar)
-        self.assertEqual(fruit['cultivar'], cultivar)
-        self.assertEqual(getattr(fruit, 'cultivar'), cultivar)
-        self.assertEqual(fruit.length_cm, length_cm)
-        self.assertEqual(fruit['length_cm'], length_cm)
-        self.assertEqual(getattr(fruit, 'length_cm'), length_cm)
-
-        # model._composed_instances is a list of the instances that were
-        # made from the anyOf/allOf/OneOf classes in model._composed_schemas
-        for composed_instance in fruit._composed_instances:
-            if composed_instance.__class__ == apple.Apple:
-                apple_instance = composed_instance
-            elif composed_instance.__class__ == banana.Banana:
-                banana_instance = composed_instance
-        self.assertEqual(
-            fruit._composed_instances,
-            [apple_instance, banana_instance]
-        )
-        # model._var_name_to_model_instances maps the variable name to the
-        # model instances which store that variable
-        self.assertEqual(
-            fruit._var_name_to_model_instances,
-            {
-                'color': [fruit],
-                'length_cm': [fruit, banana_instance],
-                'cultivar': [fruit, apple_instance],
-                'origin': [fruit, apple_instance],
-            }
-        )
-
-        # make an instance of GmFruit, a composed schema anyOf model
-        # apple test
-        color = 'red'
-        cultivar = 'golden delicious'
-        origin = 'California'
-        fruit = GmFruit(color=color, cultivar=cultivar, origin=origin)
-        # check its properties
-        self.assertEqual(fruit.color, color)
-        self.assertEqual(fruit['color'], color)
-        self.assertEqual(getattr(fruit, 'color'), color)
-        self.assertEqual(fruit.cultivar, cultivar)
-        self.assertEqual(fruit['cultivar'], cultivar)
-        self.assertEqual(getattr(fruit, 'cultivar'), cultivar)
-
-        self.assertEqual(fruit.origin, origin)
-        self.assertEqual(fruit['origin'], origin)
-        self.assertEqual(getattr(fruit, 'origin'), origin)
-
-        # check the dict representation
-        self.assertEqual(
-            fruit.to_dict(),
-            {
-                'color': color,
-                'cultivar': cultivar,
-                'origin': origin,
-            }
-        )
-
-        # model._composed_instances is a list of the instances that were
-        # made from the anyOf/allOf/OneOf classes in model._composed_schemas
-        for composed_instance in fruit._composed_instances:
-            if composed_instance.__class__ == apple.Apple:
-                apple_instance = composed_instance
-        self.assertEqual(
-            fruit._composed_instances,
-            [apple_instance]
-        )
-        # model._var_name_to_model_instances maps the variable name to the
-        # model instances which store that variable
-        self.assertEqual(
-            fruit._var_name_to_model_instances,
-            {
-                'color': [fruit],
-                'length_cm': [fruit],
-                'cultivar': [fruit, apple_instance],
-                'origin': [fruit, apple_instance],
-            }
-        )
-        # model._additional_properties_model_instances stores a list of
-        # models which have the property additional_properties_type != None
-        self.assertEqual(
-            fruit._additional_properties_model_instances, []
-        )
 
 if __name__ == '__main__':
     unittest.main()
