+++ samples/openapi3/client/petstore/python-experimental/test/test_grandparent_animal.py
@@ -14,16 +14,10 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import child_cat
-except ImportError:
-    child_cat = sys.modules[
-        'petstore_api.model.child_cat']
-try:
-    from petstore_api.model import parent_pet
-except ImportError:
-    parent_pet = sys.modules[
-        'petstore_api.model.parent_pet']
+from petstore_api.model.child_cat import ChildCat
+from petstore_api.model.parent_pet import ParentPet
+globals()['ChildCat'] = ChildCat
+globals()['ParentPet'] = ParentPet
 from petstore_api.model.grandparent_animal import GrandparentAnimal
