+++ samples/openapi3/client/petstore/python-experimental/test/test_map_test.py
@@ -14,11 +14,8 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import string_boolean_map
-except ImportError:
-    string_boolean_map = sys.modules[
-        'petstore_api.model.string_boolean_map']
+from petstore_api.model.string_boolean_map import StringBooleanMap
+globals()['StringBooleanMap'] = StringBooleanMap
 from petstore_api.model.map_test import MapTest
