+++ samples/openapi3/client/petstore/python-experimental/test/test_number_with_validations.py
@@ -28,15 +28,9 @@ class TestNumberWithValidations(unittest.TestCase):
 
     def testNumberWithValidations(self):
         """Test NumberWithValidations"""
-        valid_values = [10.0, 15.0, 20.0]
-        for valid_value in valid_values:
-            model = NumberWithValidations(valid_value)
-            assert model.value == valid_value
-
-        invalid_values = [9.0, 21.0]
-        for invalid_value in invalid_values:
-            with self.assertRaises(petstore_api.ApiValueError):
-                NumberWithValidations(invalid_value)
+        # FIXME: construct object with mandatory attributes with example values
+        # model = NumberWithValidations()  # noqa: E501
+        pass
 
 
 if __name__ == '__main__':
