+++ samples/openapi3/client/petstore/python-experimental/test/test_object_model_with_ref_props.py
@@ -14,11 +14,8 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import number_with_validations
-except ImportError:
-    number_with_validations = sys.modules[
-        'petstore_api.model.number_with_validations']
+from petstore_api.model.number_with_validations import NumberWithValidations
+globals()['NumberWithValidations'] = NumberWithValidations
 from petstore_api.model.object_model_with_ref_props import ObjectModelWithRefProps
 
 
@@ -33,15 +30,9 @@ class TestObjectModelWithRefProps(unittest.TestCase):
 
     def testObjectModelWithRefProps(self):
         """Test ObjectModelWithRefProps"""
-        from petstore_api.model.number_with_validations import NumberWithValidations
-        self.assertEqual(
-            ObjectModelWithRefProps.openapi_types,
-            {
-                'my_number': (NumberWithValidations,),
-                'my_string': (str,),
-                'my_boolean': (bool,),
-            }
-        )
+        # FIXME: construct object with mandatory attributes with example values
+        # model = ObjectModelWithRefProps()  # noqa: E501
+        pass
 
 
 if __name__ == '__main__':
