+++ samples/openapi3/client/petstore/python-experimental/test/test_parent_pet.py
@@ -14,16 +14,10 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import child_cat
-except ImportError:
-    child_cat = sys.modules[
-        'petstore_api.model.child_cat']
-try:
-    from petstore_api.model import grandparent_animal
-except ImportError:
-    grandparent_animal = sys.modules[
-        'petstore_api.model.grandparent_animal']
+from petstore_api.model.child_cat import ChildCat
+from petstore_api.model.grandparent_animal import GrandparentAnimal
+globals()['ChildCat'] = ChildCat
+globals()['GrandparentAnimal'] = GrandparentAnimal
 from petstore_api.model.parent_pet import ParentPet
 
 
@@ -38,12 +32,9 @@ class TestParentPet(unittest.TestCase):
 
     def testParentPet(self):
         """Test ParentPet"""
-
-        # test that we can make a ParentPet from a ParentPet
-        # which requires that we travel back through ParentPet's allOf descendant
-        # GrandparentAnimal, and we use the descendant's discriminator to make ParentPet
-        model = ParentPet(pet_type="ParentPet")
-        assert isinstance(model, ParentPet)
+        # FIXME: construct object with mandatory attributes with example values
+        # model = ParentPet()  # noqa: E501
+        pass
 
 
 if __name__ == '__main__':
