+++ samples/openapi3/client/petstore/python-experimental/test/test_pet.py
@@ -14,16 +14,10 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import category
-except ImportError:
-    category = sys.modules[
-        'petstore_api.model.category']
-try:
-    from petstore_api.model import tag
-except ImportError:
-    tag = sys.modules[
-        'petstore_api.model.tag']
+from petstore_api.model.category import Category
+from petstore_api.model.tag import Tag
+globals()['Category'] = Category
+globals()['Tag'] = Tag
 from petstore_api.model.pet import Pet
