+++ samples/openapi3/client/petstore/python-experimental/test/test_quadrilateral.py
@@ -14,16 +14,10 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import complex_quadrilateral
-except ImportError:
-    complex_quadrilateral = sys.modules[
-        'petstore_api.model.complex_quadrilateral']
-try:
-    from petstore_api.model import simple_quadrilateral
-except ImportError:
-    simple_quadrilateral = sys.modules[
-        'petstore_api.model.simple_quadrilateral']
+from petstore_api.model.complex_quadrilateral import ComplexQuadrilateral
+from petstore_api.model.simple_quadrilateral import SimpleQuadrilateral
+globals()['ComplexQuadrilateral'] = ComplexQuadrilateral
+globals()['SimpleQuadrilateral'] = SimpleQuadrilateral
 from petstore_api.model.quadrilateral import Quadrilateral
 
 
@@ -38,10 +32,9 @@ class TestQuadrilateral(unittest.TestCase):
 
     def testQuadrilateral(self):
         """Test Quadrilateral"""
-        instance = Quadrilateral(shape_type="Quadrilateral", quadrilateral_type="ComplexQuadrilateral")
-        assert isinstance(instance, complex_quadrilateral.ComplexQuadrilateral)
-        instance = Quadrilateral(shape_type="Quadrilateral", quadrilateral_type="SimpleQuadrilateral")
-        assert isinstance(instance, simple_quadrilateral.SimpleQuadrilateral)
+        # FIXME: construct object with mandatory attributes with example values
+        # model = Quadrilateral()  # noqa: E501
+        pass
 
 
 if __name__ == '__main__':
