+++ samples/openapi3/client/petstore/python-experimental/test/test_simple_quadrilateral.py
@@ -14,16 +14,10 @@ import sys
 import unittest
 
 import petstore_api
-try:
-    from petstore_api.model import quadrilateral_interface
-except ImportError:
-    quadrilateral_interface = sys.modules[
-        'petstore_api.model.quadrilateral_interface']
-try:
-    from petstore_api.model import shape_interface
-except ImportError:
-    shape_interface = sys.modules[
-        'petstore_api.model.shape_interface']
+from petstore_api.model.quadrilateral_interface import QuadrilateralInterface
+from petstore_api.model.shape_interface import ShapeInterface
+globals()['QuadrilateralInterface'] = QuadrilateralInterface
+globals()['ShapeInterface'] = ShapeInterface
 from petstore_api.model.simple_quadrilateral import SimpleQuadrilateral
