+++ samples/openapi3/client/petstore/python-experimental/test/test_string_enum.py
@@ -28,22 +28,9 @@ class TestStringEnum(unittest.TestCase):
 
     def testStringEnum(self):
         """Test StringEnum"""
-        inst = StringEnum(None)
-        self.assertIsNone(inst)
-
-        inst = StringEnum('approved')
-        assert isinstance(inst, StringEnum)
-
-        with self.assertRaises(petstore_api.ApiValueError):
-            StringEnum('garbage')
-
-        # make sure that we can access its allowed_values
-        assert StringEnum.allowed_values[('value',)] == {
-            'None': None,
-            'PLACED': "placed",
-            'APPROVED': "approved",
-            'DELIVERED': "delivered"
-        }
+        # FIXME: construct object with mandatory attributes with example values
+        # model = StringEnum()  # noqa: E501
+        pass
 
 
 if __name__ == '__main__':
