+++ samples/openapi3/client/petstore/python-experimental/tests_manual/test_deserialization.py
