+++ samples/openapi3/client/petstore/python-experimental/tests_manual/test_discard_unknown_properties.py
