+++ samples/openapi3/client/petstore/python-experimental/tests_manual/test_shape.py
@@ -1,0 +1,108 @@
+# coding: utf-8
+
+"""
+    OpenAPI Petstore
+
+    This spec is mainly for testing Petstore server and contains fake endpoints, models. Please do not use this for any other purpose. Special characters: \" \\  # noqa: E501
+
+    The version of the OpenAPI document: 1.0.0
+    Generated by: https://openapi-generator.tech
+"""
+
+
+import sys
+import unittest
+
+import petstore_api
+try:
+    from petstore_api.model import quadrilateral
+except ImportError:
+    quadrilateral = sys.modules[
+        'petstore_api.model.quadrilateral']
+try:
+    from petstore_api.model import triangle
+except ImportError:
+    triangle = sys.modules[
+        'petstore_api.model.triangle']
+from petstore_api.model.shape import Shape
+
+
+class TestShape(unittest.TestCase):
+    """Shape unit test stubs"""
+
+    def setUp(self):
+        pass
+
+    def tearDown(self):
+        pass
+
+    def testShape(self):
+        """Test Shape"""
+        from petstore_api.model import complex_quadrilateral
+        from petstore_api.model import simple_quadrilateral
+        from petstore_api.model import equilateral_triangle
+        from petstore_api.model import isosceles_triangle
+        from petstore_api.model import scalene_triangle
+        tri = triangle.Triangle(
+            shape_type="Triangle",
+            triangle_type="EquilateralTriangle"
+        )
+        assert isinstance(tri, equilateral_triangle.EquilateralTriangle)
+
+        tri = triangle.Triangle(
+            shape_type="Triangle",
+            triangle_type="IsoscelesTriangle"
+        )
+        assert isinstance(tri, isosceles_triangle.IsoscelesTriangle)
+
+        tri = triangle.Triangle(
+            shape_type="Triangle",
+            triangle_type="ScaleneTriangle"
+        )
+        assert isinstance(tri, scalene_triangle.ScaleneTriangle)
+
+        quad = Shape(
+            shape_type="Quadrilateral",
+            quadrilateral_type="ComplexQuadrilateral"
+        )
+        assert isinstance(quad, complex_quadrilateral.ComplexQuadrilateral)
+
+        quad = Shape(
+            shape_type="Quadrilateral",
+            quadrilateral_type="SimpleQuadrilateral"
+        )
+        assert isinstance(quad, simple_quadrilateral.SimpleQuadrilateral)
+
+        # No discriminator provided.
+        err_msg = ("Cannot deserialize input data due to missing discriminator. "
+                   "The discriminator property '{}' is missing at path: ()"
+                   )
+        with self.assertRaisesRegex(
+                petstore_api.ApiValueError,
+                err_msg.format("shapeType")
+        ):
+            Shape()
+
+        # invalid shape_type (first discriminator). 'Circle' does not exist in the model.
+        err_msg = ("Cannot deserialize input data due to invalid discriminator "
+                   "value. The OpenAPI document has no mapping for discriminator "
+                   "property '{}'='{}' at path: ()"
+                   )
+        with self.assertRaisesRegex(
+                petstore_api.ApiValueError,
+                err_msg.format("shapeType", "Circle")
+        ):
+            Shape(shape_type="Circle")
+
+        # invalid quadrilateral_type (second discriminator)
+        with self.assertRaisesRegex(
+                petstore_api.ApiValueError,
+                err_msg.format("quadrilateralType", "Triangle")
+        ):
+            Shape(
+                shape_type="Quadrilateral",
+                quadrilateral_type="Triangle"
+            )
+
+if __name__ == '__main__':
+    unittest.main()
