+++ samples/openapi3/client/petstore/python-experimental/tests_manual/util.py
